# docs/04-api/openapi.yaml
openapi: 3.0.3
info:
  title: Employee Service API
  version: 1.1.0
  description: |
    **Frozen Target API contract** — source of truth is the current runtime
    (`app/api`, `app/schemas`, `app/services`) and tests. Sibling docs:
    `api-contract-analysis.md`, `approval-api-contract.md`.

    **Target:** JWT Bearer (ADR-AUTH-JWT-01), integer business IDs + UUID User
    (ADR-ID-01), nested error envelope (ADR-ERR-03), live config without
    snapshot fields (ADR-LIVE-CFG-01), Action Engine via
    `GET .../available-actions` + `POST .../actions/{action_id}` (ADR-ACTION-01).

    Approver flow: `GET /requests/{id}` → `GET /requests/{id}/available-actions`
    → `POST /requests/{id}/actions/{action_id}`.

    `/approval-tasks/*` routes are **disabled compatibility stubs** (`409 INVALID_STATE`).
    Legacy `POST .../submit` and `POST .../cancel` are **deprecated thin aliases**
    of Action Engine (response = `RequestCard`).
...
security:
  - bearerAuth: []
...
paths:
...
  /requests:
    post:
      tags: [Requests]
      summary: Create a draft request
      description: |
        Creates a draft for an active request type. Initiator = current user.
        Creates a HistoryEvent (`create` → draft) atomically.
        Role: `employee`. Traceability: UC-04; FR-REQ-01.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestInput"
            example:
              request_type_id: 1
      responses:
        "201":
          description: Draft request created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: "`NOT_FOUND` — request type does not exist."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: "`INACTIVE_TYPE` — request type is inactive."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: "`VALIDATION` — invalid request_type_id."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Requests]
      summary: List the current employee's requests
      description: |
        Own requests only. Sorted by `created_at` descending.
        Role: `employee`. Traceability: UC-06; FR-CAB-02.
      parameters:
        - $ref: "#/components/parameters/RequestStatusFilter"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Request list items (array; pagination envelope not fixed).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequestListItem"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          description: "`VALIDATION` — unknown status or invalid page/page_size."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /requests/{request_id}:
    get:
      tags: [Requests]
      summary: Get a visible request card
      description: |
        Returns `RequestCard` when visible under ACL:
        - `employee` — own request;
        - `approver` — request with any own ApprovalTask (any task status).
        Invisible or missing request → `404 NOT_FOUND`.
        Role `admin` is not permitted on this endpoint (role gate).
        `available_actions` are **not** embedded; use
        `GET /requests/{request_id}/available-actions`.
        Form schema is **not** embedded; use `/request-types/{type_id}/schema`.
        Traceability: UC-06, UC-07; FR-REQ-04; BR-14.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Request card.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestCard"
              example:
                id: 10245
                request_type:
                  id: 1
                  code: annual_leave
                  name: Annual leave
                initiator_user_id: 33333333-3333-4333-8333-333333333333
                initiator:
                  id: 33333333-3333-4333-8333-333333333333
                  full_name: Иван Петров
                status_id: 2
                status:
                  id: 2
                  code: in_approval
                  name: На согласовании
                current_stage_id: 10
                current_stage:
                  id: 10
                  name: Line manager
                  sequence_no: 1
                created_at: "2026-09-21T09:00:00Z"
                updated_at: "2026-09-21T09:10:00Z"
                values: []
                approval_tasks:
                  - id: 501
                    stage_id: 10
                    assignee_user_id: 77777777-7777-4777-8777-777777777777
                    status: open
                    created_at: "2026-09-21T09:10:00Z"
                    completed_at: null
                comments: []
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/RequestNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Requests]
      summary: Save working values (deferred stub)
      description: |
        **Deferred stub in current runtime.** Always returns `409 INVALID_STATE`
        («сохранение полей будет доступно после следующего этапа API»).
        Not Target-active. Role gate: `employee`.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateValuesInput"
      responses:
        "409":
          description: "`INVALID_STATE` — deferred stub."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /requests/{request_id}/available-actions:
    get:
      tags: [Requests]
      summary: List available actions for a request
      description: |
        **Target (ADR-ACTION-01).** Read-only. Computes actions from live
        `ProcessTransition` for current request status and user `role_id`.
        Same visibility ACL as `GET /requests/{request_id}`.
        Does **not** filter by assignee or self-approval (enforced on execute).
        Each item is `{ id, code, name }` only — no `effect`, `target_status`.
        Roles: `employee`, `approver`.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Available actions for the current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableActionsResponse"
              example:
                available_actions:
                  - id: 3
                    code: approve
                    name: Согласовать
                  - id: 4
                    code: return
                    name: Вернуть
                  - id: 5
                    code: reject
                    name: Отклонить
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/RequestNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /requests/{request_id}/actions/{action_id}:
    post:
      tags: [Requests]
      summary: Execute a configured process action
      description: |
        **Target (ADR-ACTION-01).** Single mutation endpoint for process
        transitions. Permission and effects come from live `ProcessTransition`.
        Optional body `comment`; reject/return require non-empty trimmed comment
        → otherwise `422 VALIDATION`.
        Success returns the full updated **`RequestCard`** (not a wrapper).
        Client may re-fetch `GET .../available-actions` after success.
        Roles: `employee`, `approver`, `admin` (admin typically has no transitions).
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: action_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Action catalog id from available_actions[].id
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteActionInput"
            example:
              comment: Недостаточно обоснования
      responses:
        "200":
          description: Action applied; updated request card.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestCard"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: "`FORBIDDEN` / `FORBIDDEN_APPROVAL` (self-approval or no open assignee task)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/RequestNotFound"
        "409":
          description: |
            `REQUEST_ACTION_NOT_ALLOWED` — action not allowed for status/role;
            `TASK_DONE` — task already processed;
            `INACTIVE_TYPE`; `ROUTE_CONFIG`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: REQUEST_ACTION_NOT_ALLOWED
                  message: Действие недоступно для заявки в текущем статусе или роли
                  details: {}
        "422":
          description: "`VALIDATION` — missing required comment or submit field errors."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /requests/{request_id}/submit:
    post:
      deprecated: true
      tags: [Requests]
      summary: "[Legacy] Submit or resubmit a request"
      description: |
        **Deprecated thin alias** of Action Engine (`submit` action code).
        Target clients MUST use `POST /requests/{id}/actions/{action_id}`.
        Response = full `RequestCard` (same as execute).
        Role: `employee`. Traceability: UC-05; FR-REQ-03.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Request submitted; updated card.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestCard"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/RequestNotFound"
        "409":
          description: "`REQUEST_ACTION_NOT_ALLOWED` / `INACTIVE_TYPE` / `ROUTE_CONFIG`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: "`VALIDATION` — working values do not satisfy the live schema."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /requests/{request_id}/cancel:
    post:
      deprecated: true
      tags: [Requests]
      summary: "[Legacy] Cancel a request"
      description: |
        **Deprecated thin alias** of Action Engine (`cancel` action code).
        Target clients MUST use `POST /requests/{id}/actions/{action_id}`.
        Response = full `RequestCard`.
        Role: `employee`. Traceability: UC-10; FR-REQ-07.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Request cancelled; updated card.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestCard"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/RequestNotFound"
        "409":
          description: "`REQUEST_ACTION_NOT_ALLOWED`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /requests/{request_id}/comments:
    post:
      tags: [Requests]
      summary: Add a free initiator comment (deferred stub)
      description: |
        **Deferred stub in current runtime.** Always returns `409 INVALID_STATE`.
        Decision comments are created via Action Engine on reject/return
        (`ExecuteActionInput.comment`), not this endpoint.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentInput"
      responses:
        "409":
          description: "`INVALID_STATE` — deferred stub."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /requests/{request_id}/history:
    get:
      tags: [Requests]
      summary: Get request history
      description: |
        Returns HistoryEvent rows for a visible request (same ACL as request card).
        Envelope `{ "items": [...] }`, sorted by `at` **descending**.
        Item fields: `id`, `action`, `actor_id`, `from_state`, `to_state`, `comment`, `at`.
        Roles: `employee`, `approver`. Traceability: UC-14; FR-AUDIT-01.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: History events (newest first).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryEventListResponse"
              example:
                items:
                  - id: 703
                    action: approve
                    actor_id: 77777777-7777-4777-8777-777777777777
                    from_state: in_approval
                    to_state: approved
                    comment: null
                    at: "2026-09-21T10:00:00Z"
                  - id: 702
                    action: submit
                    actor_id: 33333333-3333-4333-8333-333333333333
                    from_state: draft
                    to_state: in_approval
                    comment: null
                    at: "2026-09-21T09:10:00Z"
                  - id: 701
                    action: create
                    actor_id: 33333333-3333-4333-8333-333333333333
                    from_state: null
                    to_state: draft
                    comment: null
                    at: "2026-09-21T09:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: "`NOT_FOUND` — request does not exist or is not visible."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Target authentication (ADR-AUTH-JWT-01).
  parameters:
...
    RequestId:
      name: request_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Request ID (integer; also the UI request number).
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    RequestStatusFilter:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/RequestStatus"
  responses:
    Unauthorized:
      description: "`UNAUTHORIZED` — missing or invalid JWT."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Требуется аутентификация
              details: {}
    Forbidden:
      description: "`FORBIDDEN` — role does not permit the operation."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: Недостаточно прав для выполнения операции
              details: {}
...
    RequestNotFound:
      description: "`NOT_FOUND` — request does not exist or is not visible to the caller."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: Объект не найден
              details: {}
    InternalError:
      description: "`INTERNAL` — unexpected server error; mutations are rolled back."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL
              message: Произошла внутренняя ошибка. Попробуйте позже
              details: {}
  schemas:
    RequestStatus:
      type: string
      enum: [draft, in_approval, returned, approved, rejected, cancelled]
...
    ErrorBody:
      type: object
      required: [code, message, details]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      description: Nested error envelope (ADR-ERR-03).
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
...
    StatusRef:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
    ActionOut:
      type: object
      required: [id, code, name]
      description: Runtime available-action item (no effect / target_status).
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
    AvailableActionsResponse:
      type: object
      required: [available_actions]
      properties:
        available_actions:
          type: array
          items:
            $ref: "#/components/schemas/ActionOut"
...
    RequestTypeRef:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
...
    CurrentStage:
      type: object
      required: [id, name, sequence_no]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        sequence_no:
          type: integer
    UserReference:
      type: object
      required: [id, full_name]
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
    FieldValue:
      type: object
      required: [field_code, value]
      properties:
        field_code:
          type: string
        value:
          type: string
          nullable: true
...
    CreateRequestInput:
      type: object
      required: [request_type_id]
      additionalProperties: false
      properties:
        request_type_id:
          type: integer
          format: int64
    CreatedRequest:
      type: object
      required:
        - id
        - request_type_id
        - initiator_user_id
        - status_id
        - status
        - current_stage_id
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        request_type_id:
          type: integer
          format: int64
        initiator_user_id:
          type: string
          format: uuid
        status_id:
          type: integer
          format: int64
        status:
          $ref: "#/components/schemas/StatusRef"
        current_stage_id:
          type: integer
          format: int64
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RequestListItem:
      type: object
      required: [id, request_type, status, current_stage, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        request_type:
          $ref: "#/components/schemas/RequestTypeRef"
        status:
          $ref: "#/components/schemas/StatusRef"
        current_stage:
          allOf:
            - $ref: "#/components/schemas/CurrentStage"
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RequestCard:
      type: object
      description: |
        Target request card (runtime). No embedded schema, available_actions,
        value_source, submit_number, or field_value_versions.
      required:
        - id
        - request_type
        - initiator_user_id
        - initiator
        - status_id
        - status
        - current_stage_id
        - created_at
        - updated_at
        - values
        - approval_tasks
        - comments
      properties:
        id:
          type: integer
          format: int64
        request_type:
          $ref: "#/components/schemas/RequestTypeRef"
        initiator_user_id:
          type: string
          format: uuid
        initiator:
          $ref: "#/components/schemas/UserReference"
        status_id:
          type: integer
          format: int64
        status:
          $ref: "#/components/schemas/StatusRef"
        current_stage_id:
          type: integer
          format: int64
          nullable: true
        current_stage:
          allOf:
            - $ref: "#/components/schemas/CurrentStage"
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        values:
          type: array
          items:
            $ref: "#/components/schemas/FieldValue"
        approval_tasks:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalTaskSummary"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    UpdateValuesInput:
      type: object
      required: [values]
      additionalProperties: false
      properties:
        values:
          type: array
          items:
            $ref: "#/components/schemas/FieldValue"
    ExecuteActionInput:
      type: object
      additionalProperties: false
      properties:
        comment:
          type: string
          nullable: true
          description: |
            Required (non-empty after trim) for reject/return; optional otherwise.
    CreateCommentInput:
      type: object
      required: [text]
      additionalProperties: false
      properties:
        text:
          type: string
...
    HistoryEventListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          description: Sorted by at descending (newest first).
          items:
            $ref: "#/components/schemas/HistoryEvent"
...